core/requirements_prod.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0
mysqlclient==2.2.4
pytz==2024.1
redis==5.0.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4